[MAC ARM Support] All tests passing with the exception of plt. Platform layer not implemented yet.

+++ src/intrinsic.h
@@ -38,7 +38,7 @@ constexpr u32 i_count_leading_zeros(TInt n) {
         return u32(__builtin_clz(u32(n)));
     }
     else {
-        return u32(__builtin_clzll(u32(n)));
+        return u32(__builtin_clzll(u64(n)));
     }
 #elif COMPILER_MSVC == 1
     if constexpr (sizeof(TInt) == 4) {
